Participant sign-in
This allows facilitators to sign in participants, creating a Session
object the first time each participant is signed in.

I had to clear out and create the migrations again to work around this
error that's caused by migrations that use
django.contrib.auth.models.User:
        https://code.djangoproject.com/ticket/22965

I'll have to drop the databases and run make install on the
staging/prod servers when this is merged.

To avoid having to do that on your local machine, you could run this
before pulling:
    ./manage.py migrate main zero