Social Support Network Map - work-in-progress
There's a lot of stuff here so I figured this would be a good place for
a review.

I'm laying out the basics for the Social Support Network Map activity -
since it requires some interactivity and data persistence I've decided to
use Ember for this. That decision may seem overkill in comparison to
Backbone, but I'm very comfortable in Ember and I think it might be a good
candidate for building these small interactives.

Unlike Weather Roulette, I'm not using the ember-cli tool to lay out and
organize the SSNM Ember application. This allows it to be more easily
integrated with the existing Django app.

My next step is to get the /ssnm/api/supporters/ endpoint working -
right now authentication is failing because DRF's SessionAuthentication
isn't recognizing an inactive user as authenticated.