Goal setting polishing + initial goal review / check-in setup.
I added some more functionality to the goal setting activity (PMT #98022)
so Jess will be able to review it. The form now populates with existing
responses and displays the correct labels for the forms (main or extra).

Also laid out some initial steps for the goal review and check-in activities
(PMT #98023).