Fix stale cohort ID bug - PMT #99554
This was solved by iterating over the form's "choices" instead of its
"queryset". "choices" is what django expects you to iterate over.

I was iterating over "queryset" as a sort of workaround, because I
need to populate the dropdown with the study ID as well as the cohort ID,
and I didn't know how to get ModelChoiceField to give me the whole object.

I added the RawModelChoiceField as a ModelChoiceField that gives the
whole instance to the template, instead of a single field.

The test for this is not quite complete since I ran into a problem where
I can't see the participant ID dropdown in the test.