YouTube video gating on VideoBlock - PMT #98025
The way I implemented this is by disabling the "next" button by
default, then enabling it with JavaScript for pages that don't
contain a video. If the page does contain a video, this link
is only enabled when that video finished playing. I'm open to
any suggestions if you think there's a better way to do this.
It seems to be working fine on the iPad so far.

Also:
* Update jQuery to 2.1.3. WORTH will only be used on iPads -- we don't
need the old IE compatibility that jquery 1.x provides.