Record per-user video views
This adds a new api endpoint at /api/watched_videos/. Whenever a user
finishes watching a video, a 'watched video' record is created. The
unlocker.js view queries these records when the user looks at any video
page, so we can then determine if the next button can be enabled or not.

This commit also contains some cleanup of the InactiveUserProfile model.