Use three.js in self-talk activity
I was spending a lot of time messing with the co-ordinates in a 2d canvas
scene, trying to make the avatar look like it's walking towards you down a
road. I realized this is actually 3d scene, so I finally had an excuse to
learn three.js. Now it's straightforward to implement the perspective without
having to fake the math through trial and error.

I've been testing this on the iPad and it looks like it will work fine there.