Participant sign-in fixes, preliminary work for PMT #100768
I've fixed tests when using postgresql as test db, also, I'm now using
postgresql on travis, just like we're doing in PMT. This is because
there is some code that uses a postgresql-specific .distinct() method
that we need to be testing.

This commit also removes an unnecessary set_password() call in the
participant sign-in process.