Participant sign-in form updates
* Removed my non-standard RawModelChoiceField, RawModelChoiceIterator
form overrides, instead getting this functionality with the
get_participant templatetag.

* This includes a performance improvement to
Participant.highest_module_accessed, to improve the sluggish
loading time on the sign-in page. I was able to remove all
the test skips as a side effect, since this method no longer
uses `.distinct()`.

* Sign-in form now saves selected participant and destination
when displaying a validation error (PMT #100762)

Addressing PMTs #100796, #100797, #100762