Add caching
This adds caching to worth as described here:
https://docs.djangoproject.com/en/1.8/topics/cache/

For local development, this will use the LocMemCache, which requires no
setup for developers:
https://docs.djangoproject.com/en/1.8/topics/cache/#local-memory-caching

Optionally, you can still override this in local_settings.py to use a local
memcached instance.

On staging and production, worth will connect to memcached. I've added
memcached to scuba and django in salt.