Simplify sign-in form destination picker
This removes a feature that had a confusing UI, and was
just as confusing to program. Instead of a complex
radio button / dropdown combination, the participant's
destination is now explicitly chosen as Session 1 through
5. The participant's current progress, as a percentage, is
displayed next to each session. We can improve the styles
here, but the basic functionality is done.

PMTs #101078, #101095