Integration testing for the intervention - PMT #100729
This adds some integration testing to the avatar activity and
one of the quizzes. I'm also building out a mock instance of the
worth application as "WorthModuleFactory" that's used both in
the behave tests and the unit tests.

For now, in behave, I still have to go through the participant
sign-in process manually. I couldn't get the
create_pre_authenticated_session scheme to work for signing in
participants. There's comments on this in the code.