Remove duplicate success messages - PMT #101324
Because the goal setter renders its own submit button, it's
difficult to put the global success message in the same area as it.

It's possible of course, but for now I think it would just be more
trouble than it's worth, because it looks fine like this (see
attached screenshot).