Behave fixes
One of the "Then" functions in the bdd steps didn't have an `assert`
on it, so it wasn't doing anything. This uncovered a bug with the
success alert on the goal setting page, and some problems with the
behave tests, which I've fixed.

I haven't gotten the create_pre_authenticated_session stuff to work,
even for normal, non-participant users, but doing the manual
sign-in is working for now.