Implement new study ID format - PMT #104086
This change is fine to be pushed whenever: the old participants (there
are 3 on production) will stay on the system, and we can update their
replacement IDs when we have them. I had told Jess we would need to
manually make a migration to handle this, and that we need the
replacement IDs in order to make this change, but it turns out that's
not the case.