:arrow_up: selenium 2.48.0 + remove splinter
Selenium 2.48.0 is required for FF 43+, which is the version of Firefox
I have on my computer.

I'm experimenting with trying to fix the unreliable "Manage Participant
IDs" step in behave. I've moved some things around, added a fix for
link clicking. I'm calling selenium methods directly now instead of going
through splinter. This is to eliminate a potentially problematic layer of
abstraction.