Render next button as enabled by default - PMT #104530
A participant encountered a bug where they couldn't continue past
the SSNM activity in session 4, even after signing out and signing
back in to WORTH. I was unable to reproduce the bug on my iPad,
using the production database on staging.

The way I implemented the locking feature was: render the "next" button
as disabled, then unlock it with javascript if the right conditions are
met. I'm not sure why I did it this way, because it relies on all the
javascript getting loaded and working correctly in order to proceed past
each pageblock.

I think it will be more reliable to render the next button as 'enabled'
by default, and disable it with javascript given certain conditions.

Then, if there's any issue with the javascript, the worst that can
happen is the user is able to proceed to the next page when they
shouldn't be allowed to. Not nearly as bad as preventing a user from
proceeding.