Remove invalid attribute on sign in form
The participant sign in form had `action="""`, which is invalid
HTML. The action attribute isn't necessary for this form so I've
removed it.

I also moved the csrf_token up to the `<form>` tag. This makes it
obvious that this form has a CSRF token when grepping through the
code.