add django-pagetree-epub
That library is an experimental feature that started out in forest. I
pulled it out into its own module so we can use it here.

It builds a very basic epub book out of a pagetree site.

As you will probably see, it is very, very barebones at the moment, only
supporting a couple of basic blocks. We never really got very far on
that in Forest because we didn't actually having anything driving it.

Especially since WORTH is mostly video and interactive content, the epub
that it currently generates is laughably unusable, with 'Unrenderable
Block' for most of the content.

But, we have to start somewhere. With this in place, we can start
working block by block to figure out how it should be rendered in an
epub, or to replace it outright with something else.