remove study_id_validator()
function does nothing. Looks like it was once used for doing some
validation on a migration, but has since been emptied out. No point in
keeping it around. Removed the references to it from
migrations. Obviously, if those migrations run, they won't do whatever
validation the function used to do, but that's been true ever since it
was turned into a no-op.